Computer, Program, Software Service Tests

See merge request Matthew.Chanco/group-project-sydney!70

+++ src/main/java/com/fdmgroup/PCTrack/service/ProgramService.java
@@ -11,12 +11,12 @@ import com.fdmgroup.PCTrack.model.Software;
 
 @Service
 public class ProgramService {
-	private ProgramRepository softwareRepository;
+	private SoftwareRepository softwareRepository;
 	private ProgramRepository programRepository;
 
 	public ProgramService(ProgramRepository programRepository, SoftwareRepository softwareRepository) {
 		super();
-		this.softwareRepository = programRepository;
+		this.softwareRepository = softwareRepository;
 		this.programRepository = programRepository;
 	}
 	
@@ -67,13 +67,14 @@ public class ProgramService {
 
 	public void deleteBySoftwareId(int softwareId)
 	{
-		for (Program p : findAllPrograms()) {
+		List<Program> programs = findAllPrograms();
+		if (programs.isEmpty()){
+			throw new RuntimeException("Program does not exist");
+		}
+		for (Program p : programs) {
 			int index = p.getSoftware().getSoftwareId();
 			if(index == softwareId) {
 				programRepository.deleteById(softwareId);
-			}
-			else {
-				throw new RuntimeException("Program does not exist");
 			}
 		}
 	}
